migration_deploy_simple/requirements.txt
functions-framework==3.*
google-cloud-storage==2.16.0
